+++ implementation/src/main/java/com/avioconsulting/mule/testing/dsl/invokers/FlowRunnerImpl.groovy
@@ -4,7 +4,10 @@ import com.avioconsulting.mule.testing.InvokerEventFactory
 import com.avioconsulting.mule.testing.muleinterfaces.RuntimeBridgeTestSide
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.EventWrapper
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.FlowWrapper
+import groovy.util.logging.Log4j2
+import org.apache.logging.log4j.CloseableThreadContext
 
+@Log4j2
 class FlowRunnerImpl implements
         FlowRunner,
         BatchRunner {
@@ -16,6 +19,7 @@ class FlowRunnerImpl implements
     private final InvokerEventFactory invokerEventFactory
     private final FlowWrapper flow
     private final String flowName
+    private final CloseableThreadContext.Instance threadContext
 
     FlowRunnerImpl(RuntimeBridgeTestSide runtimeBridge,
                    FlowWrapper flowMessageProcessor,
@@ -24,6 +28,8 @@ class FlowRunnerImpl implements
         this.flow = flowMessageProcessor
         this.runtimeBridge = runtimeBridge
         this.invokerEventFactory = runtimeBridge
+        threadContext = CloseableThreadContext.push('Flow invocation')
+        this.threadContext.put('flowInvocation', flowName)
     }
 
     def json(@DelegatesTo(JsonInvoker) Closure closure) {
@@ -73,6 +79,7 @@ class FlowRunnerImpl implements
 
     EventWrapper getEvent() {
         assert invoker: 'Need to specify a proper format! (e.g. json)'
+        log.info 'Creating event to invoke flow'
         def code = closure.rehydrate(invoker,
                                      this,
                                      this)
@@ -97,5 +104,10 @@ class FlowRunnerImpl implements
         }
 
         response
+    }
+
+    def closeLogContext() {
+        log.info 'Completed flow execution'
+        threadContext.close()
     }
 }