+++ implementation/src/main/java/com/avioconsulting/mule/testing/transformers/json/input/JacksonInputTransformer.groovy
@@ -2,7 +2,9 @@ package com.avioconsulting.mule.testing.transformers.json.input
 
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.ConnectorInfo
 import com.fasterxml.jackson.databind.ObjectMapper
+import groovy.util.logging.Log4j2
 
+@Log4j2
 class JacksonInputTransformer<T extends ConnectorInfo> extends
         Common<T> {
     def mapper = new ObjectMapper()
@@ -18,12 +20,17 @@ class JacksonInputTransformer<T extends ConnectorInfo> extends
 
     def transform(String jsonString) {
         if (jsonString == '' || jsonString == null) {
+            log.info 'No JSON to unmarshal'
             return null
         }
         def errors = []
         for (def klass : inputClasses) {
             try {
-                return mapper.readValue(jsonString, klass)
+                log.info 'Unmarshalling JSON {} to klass {}',
+                         jsonString,
+                         klass
+                return mapper.readValue(jsonString,
+                                        klass)
             }
             catch (e) {
                 errors << e